+++ packages/insomnia/src/ui/components/settings/import-export.tsx
@@ -1,7 +1,6 @@
 import { format } from 'date-fns';
 import { getProductName } from 'insomnia/src/common/constants';
 import { getWorkspaceLabel } from 'insomnia/src/common/get-workspace-label';
-import { exportRequestsHAR, exportWorkspacesHAR } from 'insomnia/src/common/har';
 import { getInsomniaV5DataExport } from 'insomnia/src/common/insomnia-v5';
 import { isNotNullOrUndefined } from 'insomnia/src/common/misc';
 import { AnalyticsEvent } from 'insomnia/src/ui/analytics';
@@ -170,6 +169,7 @@ export const exportProjectToFile = (activeProjectName: string, workspacesForActi
             if (!fileName) {
               return;
             }
+            const { exportWorkspacesHAR } = await import('insomnia/src/common/har');
             const stringifiedExport = await exportWorkspacesHAR(
               workspacesForActiveProject,
               shouldExportPrivateEnvironments,
@@ -332,6 +332,7 @@ export const exportRequestsToFile = (workspaceId: string, requestIds: string[]) 
       try {
         switch (selectedFormat) {
           case VALUE_HAR: {
+            const { exportRequestsHAR } = await import('insomnia/src/common/har');
             stringifiedExport = await exportRequestsHAR(requests, shouldExportPrivateEnvironments);
             break;
           }
