+++ packages/insomnia/src/utils/vault.test.ts
@@ -1,12 +1,11 @@
 // @vitest-environment jsdom
 import { beforeEach, describe, expect, it, vi } from 'vitest';
 
+import { decryptSecretValue, encryptSecretValue } from './vault-crypto';
 import {
   base64decode,
   base64encode,
-  decryptSecretValue,
   decryptVaultKeyFromSession,
-  encryptSecretValue,
 } from './vault';
 
 vi.mock('../models/settings', () => ({
