+++ packages/insomnia/src/utils/vault.ts
@@ -1,6 +1,5 @@
 import { services } from '~/insomnia-data';
 
-import { type AESMessage, decryptAES, encryptAES } from '../account/crypt';
 import { getInsomniaVaultKey, PLAYWRIGHT_TEST } from '../common/constants';
 
 export const base64encode = (input: string | JsonWebKey) => {
@@ -60,33 +59,4 @@ export const getVaultKeyFromStorage = async (accountId: string) => {
 
 export const deleteVaultKeyFromStorage = async (accountId: string) => {
   await window.main.secretStorage.deleteSecret(getVaultSecretKey(accountId));
-};
-
-export const encryptSecretValue = (rawValue: string, symmetricKey: JsonWebKey) => {
-  if (typeof symmetricKey !== 'object' || Object.keys(symmetricKey).length === 0) {
-    // invalid symmetricKey
-    return rawValue;
-  }
-  try {
-    const encryptResult = encryptAES(symmetricKey, rawValue);
-    const encryptedValue = base64encode(encryptResult);
-    return encryptedValue;
-  } catch {
-    // return original value if encryption fails
-    return rawValue;
-  }
-};
-
-export const decryptSecretValue = (encryptedValue: string, symmetricKey: JsonWebKey) => {
-  if (typeof symmetricKey !== 'object' || Object.keys(symmetricKey).length === 0) {
-    // invalid symmetricKey
-    return encryptedValue;
-  }
-  try {
-    const jsonWebKey = base64decode(encryptedValue, true) as AESMessage;
-    return decryptAES(symmetricKey, jsonWebKey);
-  } catch {
-    // return origin value if failed to decrypt
-    return encryptedValue;
-  }
 };
